Add super hacky quick page directory, please forgive me
