+++ Controllers/Umbraco/BaseUmbracoAuthorisedController.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Text.Json;
+using Umbraco.Cms.Web.BackOffice.Controllers;
+
+namespace AccessibilityReporter.Controllers.Umbraco
+{
+    public class BaseUmbracoAuthorisedController : UmbracoAuthorizedApiController
+    {
+        protected JsonResult Json(object payload)
+			=> new JsonResult(payload, new JsonSerializerOptions
+			{
+				PropertyNamingPolicy = JsonNamingPolicy.CamelCase
+			});
+	}
+}
+
