+++ Controllers/Umbraco/ConfigController.cs
@@ -1,0 +1,19 @@
+﻿using AccessibilityReporter.Core.Interfaces;
+using Microsoft.AspNetCore.Mvc;
+
+namespace AccessibilityReporter.Controllers.Umbraco
+{
+    public class ConfigController : BaseUmbracoAuthorisedController
+	{
+        private readonly IAccessibilityReporterSettings _settings;
+
+        public ConfigController(IAccessibilityReporterSettings settings)
+        {
+            _settings = settings;
+        }
+
+        [HttpGet]
+        public JsonResult Current()
+            => Json(_settings);
+    }
+}
