+++ Infrastructure/Config/AccessibilityReporterSettingsFactory.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using AccessibilityReporter.Core.Interfaces;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace AccessibilityReporter.Infrastructure.Config
